requirements.txt
asgiref==3.4.1
astroid==2.8.0
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
pylint==2.11.1
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
